Refactor[submodule] #3: Initialize controller & validator Profile
