+++ devconnector/routes/apis/profile.js
@@ -1,182 +1,51 @@
 const express = require('express')
 const router = express.Router()
-const { check, validationResult } = require('express-validator')
 
+// Middleware
 const auth = require('../../middleware/auth')
 
-const Profile = require('../../models/Profile')
+// Controller
+const {
+  getCurrentUserProfile,
+  createProfile,
+  getProfiles,
+  getProfileByUserId,
+  updateExperience
+} = require('../../controllers/profile')
+
+// Validator
+const {
+  createProfileValidator,
+  updateExperienceValidator
+} = require('../../validators/profile')
 
 /**
  * @route GET api/profile/me
  * @desc  Get current user profile
  * @access  Private
  */
-router.get('/me', auth, async (req, res) => {
-  try {
-    const profile = await Profile
-      .findOne({ user: req.user.id })
-      .populate('user', ['name', 'avatar'])
-
-    if (!profile) {
-      return res
-        .status(400)
-        .json({
-          msg: 'There is no profile for this user.'
-        })
-    }
-
-    res.json({ profile })
-  } catch (error) {
-    res.status(500).send({
-      msg: 'Server Error.'
-    })
-  }
-})
+router.get('/me', auth, getCurrentUserProfile)
 
 /**
  * @route POST api/profile
  * @desc  Create / Update user profile
  * @access  Private
  */
-router.post(
-  '/',
-  [
-    auth,
-    [
-      check('status', 'Status is required.').not().isEmpty(),
-      check('skills', 'Skills is required.').not().isEmpty()
-    ]
-  ],
-  async (req, res) => {
-    const errors = validationResult(req)
-
-    if (!errors.isEmpty()) {
-      return res
-        .status(400)
-        .json({ errors: errors.array() })
-    }
-
-    const {
-      company,
-      website,
-      location,
-      status,
-      skills,
-      bio,
-      githubProfile,
-      experience,
-      education,
-      youtube,
-      twitter,
-      facebook,
-      linkedin,
-      instagram,
-      devto
-    } = req.body
-
-    // Build profile object
-    const profileFields = {}
-    profileFields.user = req.user.id
-
-    if (company) profileFields.company = company
-    if (website) profileFields.website = website
-    if (location) profileFields.location = location
-    if (status) profileFields.status = status
-    if (bio) profileFields.bio = bio
-    if (githubProfile) profileFields.githubProfile = githubProfile
-    if (experience) profileFields.experience = experience
-    if (education) profileFields.education = education
-    if (skills) {
-      profileFields.skills = skills.split(',').map(skill => skill.trim())
-    }
-
-    profileFields.social = {}
-    if (youtube) profileFields.social.youtube = youtube
-    if (twitter) profile.social.twitter = twitter
-    if (facebook) profileFields.social.facebook = facebook
-    if (linkedin) profileFields.social.linkedin = linkedin
-    if (instagram) profileFields.social.instagram = instagram
-    if (devto) profileFields.social.devto = devto
-
-    try {
-      let profile = await Profile.findOne({ user: req.user.id })
-
-      if (profile) {
-        // Update
-        profile = await Profile.findOneAndUpdate(
-          { user: req.user.id },
-          { $set: profileFields },
-          { new: true }
-        )
-
-        return res.json(profile)
-      }
-
-      // Create
-      profile = new Profile(profileFields)
-      await profile.save()
-      res.json(profile)
-    } catch (error) {
-      res.status(500).send({
-        msg: 'Server Error.'
-      })
-    }
-  }
-)
+router.post('/', createProfileValidator, createProfile)
 
 /**
  * @route GET api/profile
  * @desc  Get all user profiles
  * @access  Public
  */
-router.get('/', async (req, res) => {
-  try {
-    const profiles = await Profile
-      .find()
-      .populate('user', ['name', 'avatar'])
-
-    res.json(profiles)
-  } catch (error) {
-    res.status(500).send({
-      msg: 'Server Error.'
-    })
-  }
-})
+router.get('/', getProfiles)
 
 /**
  * @route GET api/profile/user/:user_id
  * @desc  Get profile by user ID
  * @access  Public
  */
-router.get('/user/:user_id', async (req, res) => {
-  try {
-    const profile = await Profile
-      .findOne({ user: req.params.user_id })
-      .populate('user', ['name', 'avatar'])
-
-    if (!profile) {
-      return res
-        .status(400)
-        .json({
-          msg: 'Profile not found.'
-        })
-    }
-
-    res.json(profile)
-  } catch (error) {
-    if (error.kind === 'ObjectId') {
-      return res
-        .status(400)
-        .json({
-          msg: 'Profile not found.'
-        })
-    }
-
-    res.status(500).send({
-      msg: 'Server Error.'
-    })
-  }
-})
+router.get('/user/:user_id', getProfileByUserId)
 
 /**
  * @route PATCH api/profile/update/experience
@@ -185,53 +54,8 @@ router.get('/user/:user_id', async (req, res) => {
  */
 router.patch(
   '/update/experience',
-  [
-    auth,
-    check('title', 'Title is required.').not().isEmpty(),
-    check('company', 'Company is required.').not().isEmpty(),
-    check('from', 'From date is required.').not().isEmpty()
-  ],
-  async (req, res) => {
-    const errors = validationResult(req)
-
-    if (!errors.isEmpty()) {
-      return res
-        .status(400)
-        .json({ errors: errors.array() })
-    }
-
-    const {
-      title,
-      company,
-      location,
-      from,
-      to,
-      current,
-      description
-    } = req.body
-
-    const newExperience = {
-      title,
-      company,
-      location,
-      from,
-      to,
-      current,
-      description
-    }
-
-    try {
-      const profile = await Profile.findOne({ user: req.user.id })
-      profile.experience.unshift(newExperience)
-      profile.save()
-
-      res.json(profile)
-    } catch (error) {
-      res.status(500).send({
-        msg: 'Server Error.'
-      })
-    }
-  }
+  updateExperienceValidator,
+  updateExperience
 )
 
 module.exports = router
